requirements.txt
qrcode[pil]
Pillow
reportlab
